# Optimize ST operations